fix: action infos
