+++ packages/genome-spy/src/view/sampleView/sampleSlice.js
@@ -13,6 +13,17 @@ import {
     wrapAccessorForComparison,
 } from "./sampleOperations";
 
+import { format as d3format } from "d3-format";
+import { html } from "lit";
+import {
+    faSortAmountDown,
+    faFilter,
+    faMedal,
+    faObjectGroup,
+    faCircle,
+    faTrashAlt,
+} from "@fortawesome/free-solid-svg-icons";
+
 /**
  * @typedef {import("./sampleState").SampleHierarchy} SampleHierarchy
  * @typedef {import("./sampleState").Group} Group
@@ -26,6 +37,18 @@ import {
  * @template P
  * @typedef {import("@reduxjs/toolkit").PayloadAction<P>} PayloadAction
  */
+
+const SORT_BY_NAME = "sortByName";
+const SORT_BY = "sortBy";
+const RETAIN_FIRST_OF_EACH = "retainFirstOfEach";
+const FILTER_BY_NOMINAL = "filterByNominal";
+const FILTER_BY_QUANTITATIVE = "filterByQuantitative";
+const REMOVE_UNDEFINED = "removeUndefined";
+//const REMOVE_BY_ID = "removeById";
+const GROUP_BY_NOMINAL = "groupByNominal";
+const GROUP_TO_QUARTILES = "groupToQuartiles";
+
+const SLICE_NAME = "sampleView";
 
 /**
  * @returns {SampleHierarchy}
@@ -59,7 +82,7 @@ export function createSampleSlice(getAttributeInfo) {
     };
 
     return createSlice({
-        name: "sampleView",
+        name: SLICE_NAME,
         initialState: createInitialState(),
         reducers: {
             setSamples: (
@@ -111,7 +134,7 @@ export function createSampleSlice(getAttributeInfo) {
                 };
             },
 
-            sortBy: (
+            [SORT_BY]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").SortBy>} */ action
             ) => {
@@ -127,11 +150,11 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            sortByName: (state) => {
-                // TODO
+            [SORT_BY_NAME]: (state) => {
+                alert("TODO");
             },
 
-            retainFirstOfEach: (
+            [RETAIN_FIRST_OF_EACH]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").SortBy>} */
                 action
@@ -144,7 +167,7 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            filterByQuantitative: (
+            [FILTER_BY_QUANTITATIVE]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").FilterByQuantitative>} */
                 action
@@ -159,7 +182,7 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            filterByNominal: (
+            [FILTER_BY_NOMINAL]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").FilterByNominal>} */
                 action
@@ -174,7 +197,7 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            removeUndefined: (
+            [REMOVE_UNDEFINED]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").RemoveUndefined>} */
                 action
@@ -184,7 +207,7 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            groupByNominal: (
+            [GROUP_BY_NOMINAL]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").GroupByNominal>} */
                 action
@@ -197,7 +220,7 @@ export function createSampleSlice(getAttributeInfo) {
                 );
             },
 
-            groupToQuartiles: (
+            [GROUP_TO_QUARTILES]: (
                 state,
                 /** @type {PayloadAction<import("./payloadTypes").GroupToQuartiles>} */
                 action
@@ -252,7 +275,7 @@ function getSampleGroups(sampleHierarchy) {
  * @returns {SampleHierarchy}
  */
 export function sampleHierarchySelector(state) {
-    return state.sampleView;
+    return state[SLICE_NAME];
 }
 
 /**
@@ -321,5 +344,137 @@ export function* iterateGroupHierarcy(group) {
                 yield [group, ...elem];
             }
         }
+    }
+}
+
+const attributeNumberFormat = d3format(".4");
+
+/** @type {Record<string, string>} */
+const verboseOps = {
+    lt: "less than",
+    lte: "less than or equal to",
+    eq: "equal to",
+    gte: "greater than or equal to",
+    gt: "greater than",
+};
+
+/**
+ * Describes an action for displaying it in menus or provenance tracking.
+ *
+ * @param {import("@reduxjs/toolkit").PayloadAction<any>} action
+ * @param {import("./compositeAttributeInfoSource").AttributeInfoSource} getAttributeInfo
+ * @returns {import("../../app/provenance").ActionInfo}
+ */
+export function getActionInfo(action, getAttributeInfo) {
+    if (!action.type.startsWith(SLICE_NAME)) {
+        return;
+    }
+
+    // It would be great to have working payload typings here
+    const payload = action.payload;
+
+    const attributeInfo = getAttributeInfo(payload.attribute);
+    const attributeName = attributeInfo?.name;
+    const attributeTitle =
+        attributeInfo?.title || html` <em>${attributeName}</em> `;
+
+    const template = {
+        attributeName, // TODO: This may actually be unnecessary
+    };
+
+    const actionType = action.type.substring(SLICE_NAME.length + 1);
+
+    switch (actionType) {
+        case SORT_BY_NAME:
+            return {
+                ...template,
+                title: "Sort by sample name",
+                icon: faSortAmountDown,
+            };
+        case SORT_BY:
+            return {
+                ...template,
+                title: "Sort by",
+                provenanceTitle: html` Sort by ${attributeTitle} `,
+                icon: faSortAmountDown,
+            };
+        case RETAIN_FIRST_OF_EACH:
+            return {
+                ...template,
+                title: html`
+                    Retain first sample of each unique
+                    <em>${attributeName}</em>
+                `,
+                provenanceTitle: html`
+                    Retain first sample of each unique ${attributeTitle}
+                `,
+
+                icon: faMedal,
+            };
+        case FILTER_BY_NOMINAL: {
+            /** @param {string | import("lit").TemplateResult} attr */
+            const makeTitle = (attr) => html`
+                ${payload.action == "remove" ? "Remove" : "Retain"} samples
+                having
+                ${payload.values[0] === undefined || payload.values[0] === null
+                    ? html` undefined ${attr} `
+                    : html`
+                          ${attr} =
+                          <strong>${payload.values[0]}</strong>
+                      `}
+            `;
+
+            return {
+                ...template,
+                title: makeTitle(html` <em>${attributeName}</em> `),
+                provenanceTitle: makeTitle(attributeTitle),
+                icon: payload.action == "remove" ? faTrashAlt : faFilter,
+            };
+        }
+        case FILTER_BY_QUANTITATIVE: {
+            /** @param {string | import("lit").TemplateResult} attr */
+            const makeTitle = (attr) => html`
+                Retain samples having ${attr} ${verboseOps[payload.operator]}
+                <strong>${attributeNumberFormat(payload.operand)}</strong>
+            `;
+
+            return {
+                ...template,
+                title: makeTitle(html` <em>${attributeName}</em> `),
+                provenanceTitle: makeTitle(attributeTitle),
+                icon: faFilter,
+            };
+        }
+        case REMOVE_UNDEFINED:
+            return {
+                ...template,
+                title: "Remove samples having missing attribute",
+                provenanceTitle: html`
+                    Remove samples having missing ${attributeTitle}
+                `,
+                icon: faFilter,
+            };
+        case GROUP_BY_NOMINAL:
+            return {
+                ...template,
+                title: "Group by",
+                provenanceTitle: html` Group by ${attributeTitle} `,
+                icon: faObjectGroup,
+            };
+        case GROUP_TO_QUARTILES:
+            return {
+                ...template,
+                title: "Group to quartiles",
+                provenanceTitle: html`
+                    Group to quartiles by ${attributeTitle}
+                `,
+                icon: faObjectGroup,
+            };
+        default:
+            return {
+                ...template,
+                title: JSON.stringify(action),
+                icon: faCircle,
+            };
     }
 }
